% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_regsim_method.R
\name{print.regsim}
\alias{print.regsim}
\title{Print method for class regsim}
\usage{
\method{print}{regsim}(x, ...)
}
\arguments{
\item{x}{an object of class regsim to be printed}

\item{...}{additional arguments to be passed to \code{print}}
}
\description{
Print method for class regsim
}
