% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regsim_glm.R
\name{regsim.glm}
\alias{regsim.glm}
\title{Regression Simulation for Generalized Linear Models}
\usage{
\method{regsim}{glm}(object, x, num = 1000, link = logit, ...)
}
\arguments{
\item{object}{a regression model, usually, a result of a call to glm}

\item{x}{a list of explanatory variables}

\item{num}{number of iterations to run}

\item{link}{model link function}

\item{...}{additional arguments passed to class-specific functions}
}
\description{
\code{regsim.glm} simulates quantities of interest from a generalized linear model
}
\examples{
library(regsim)

y <- swiss$Fertility > median(swiss$Fertility)
model <- glm(y ~ Education + Agriculture, family = "binomial", data = swiss)
x <- list(
  Agriculture = seq(1, 100, 5),
  Education = mean(swiss$Education)
)
sim <- regsim(model, x)
summary(sim)
}
