% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regsim_lm.R
\name{regsim.lm}
\alias{regsim.lm}
\title{Regression Simulation for Linear Models}
\usage{
\method{regsim}{lm}(object, x, num = 1000, ...)
}
\arguments{
\item{object}{a regression model, usually, a result of a call to \link[stats]{lm}}

\item{x}{a list of explanatory variables}

\item{num}{number of iterations to run}

\item{...}{additional arguments passed to class-specific functions}
}
\description{
\code{regsim.lm} simulate quantities of interest from a linear regression model
}
\examples{
library(regsim)

model <- lm(mpg ~ wt + cyl, data = mtcars)
x <- list(wt = seq(1, 5, 0.1))
sim <- regsim(model, x)
summary(sim)
}
