% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.regsim}
\alias{summary.regsim}
\title{Summarizing Regression Simulations}
\usage{
\method{summary}{regsim}(object, detail = TRUE, long = detail, ...)
}
\arguments{
\item{object}{an object of class \code{regsim}, usually obtained by calling the
\link{regsim} function.}

\item{detail}{print detail format}

\item{long}{print summary in long format (default is TRUE in detail mode)}

\item{...}{additional arguments passed to class-specific functions}
}
\description{
summary method for class "regsim"
}
\examples{
library(regsim)

model <- lm(mpg ~ wt + cyl, data = mtcars)
x <- list(wt = seq(1, 5, 0.1))
sim <- regsim(model, x)
summary(sim)
}
