% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.regsim}
\alias{plot.regsim}
\title{Plot Regression Simulations}
\usage{
\method{plot}{regsim}(x, formula, ...)
}
\arguments{
\item{x}{an object of class \code{regsim}, usually obtained by calling the
\link{regsim} function.}

\item{formula}{a \link[stats]{formula} such as ~x or ~x1 + x2}

\item{...}{additional arguments passed to class-specific functions}
}
\description{
plots regression simulation results
}
\examples{
library(regsim)

model <- lm(mpg ~ wt + cyl, data = mtcars)
x <- list(
  wt = seq(1, 5, 0.1),
  cyl = mean(mtcars$cyl)
)
sim <- regsim(model, x)
plot(sim, ~wt)
}
