% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.regsim}
\alias{plot.regsim}
\title{Plot Regression Simulations}
\usage{
plot.regsim(x, var, ...)
}
\arguments{
\item{x}{an object of class "regsim", usually obtained by calling the
\code{regsim} function.}

\item{var}{variable to plot on the x-axis}

\item{...}{additional arguments passed to class-specific functions}
}
\description{
plots regression simulation results
}
\examples{
library(regsim)

model <- lm(mpg ~ wt + cyl, data = mtcars)
sim <- regsim(model, list(wt = seq(1, 5, 0.1), cyl = mean(mtcars$cyl)))
plot(sim, ~wt)
}
