% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regsim_method.R
\name{plot.regsim}
\alias{plot.regsim}
\title{Plot method for class regsim}
\usage{
\method{plot}{regsim}(x, sizeMan = 10, edge.label.cex = 1.5,
  fixedStyle = 1, nCharNodes = 0, ...)
}
\arguments{
\item{x}{an object of class regsim to be printed}

\item{sizeMan}{numeric; controls scale of the variables and boxes, defaults to 10}

\item{edge.label.cex}{numeric; controls scale of the path coefficient. defaults to 1.5}

\item{fixedStyle}{numeric or vector of an lty and a color for the arrows. defaults to 1 for black, solid lines}

\item{...}{additional arguments to be passed to \code{semPaths}}
}
\description{
Plot method for class regsim
}
