% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regsim.R
\name{regsim}
\alias{regsim}
\title{Regression Simulation}
\usage{
regsim(object, x, num = 1000, ...)
}
\arguments{
\item{object}{a regression model, usually, a result of a call to
\link[stats]{lm} or \link[stats]{glm}}

\item{x}{a list of explanatory variables}

\item{num}{number of iterations to run}

\item{...}{additional arguments passed to class-specific functions}
}
\description{
\code{regsim} simulates quantities of interest from regression models
}
\examples{
library(regsim)

model <- lm(mpg ~ wt + cyl, data = mtcars)
x <- list(
  wt = seq(1, 5, 0.1),
  cyl = mean(mtcars$cyl)
)
sim <- regsim(model, x)
summary(sim)
}
