% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regsim_function.R
\name{regsim}
\alias{regsim}
\title{Fit and appraise regression models using simulated data}
\usage{
regsim(reps, n, true.model, fit.model, targetparm, targetval,
  interval = 0.95, ...)
}
\arguments{
\item{reps}{the number of repetitions to perform}

\item{n}{the sample size of the generated data}

\item{true.model}{character; the population model used to generate the data specified in \code{lavaan} syntax. See example.}

\item{fit.model}{character; the model to the fitted to the data in \code{lm()} model formula syntax}

\item{targetparm}{character; the focal predictor. Must be included in the model formula passed to \code{true.model}}

\item{targetval}{the true value of the path coefficient relating the focal predictor to the response variable}

\item{interval}{the confidence interval width, defaults to 0.95.}

\item{...}{additional arguments passed to lavaan's \code{simulateData} function. For example, one could invoke the
\code{kurtosis} or \code{skewness} arguments to generate non-normal data}
}
\value{
A list containing the following:
 \item{true.model}{the true model from which data are simulated}
 \item{fit.model}{the regression model that is fit to the data}
 \item{b}{a vector of target parameter estimates across the repetitions}
 \item{data}{a data frame containing the simulated data from the first repetition}
 \item{true.DAG}{the true.model expressed as an object of class dagitty}
 \item{targetval}{the true value of the target parameter, used to calculate bias, RMSE, and coverage}
 \item{expected.b}{the mean value of the target parameter across the repetitions}
 \item{bias}{the mean difference between the estimated and target values of the
 parameter across repetitions}
 \item{empirical.CI}{the empirical confidence interval boundaries across the repetitions}
 \item{coverage}{the proportion of repetitions in which the confidence interval captured
 the specified target value of the parameter. This should appoximately equal
 the specified confidence level}
 \item{empirical.SE}{the standard deviation of the parameter estimates}
 \item{analytic.SE}{the mean of the analytic standard errors across repetitions}
 \item{RMSE}{the root mean squared error}
 \item{adjustment.sets}{the sets of covariates that must be included in the fitted model
   to yield an unbiased and consistent estimate of the effect of the target parameter on the
   response variable in fit.model}
}
\description{
This function provides a quick and convenient means for simulating data from a
path diagram or SEM, fitting a regression model to the simulated data, and appraising the
model's performance with respect to a particular target parameter and its standard
error. The function calculates the bias, coverage rate, RMSE, and analytical and empirical
standard errors.
}
\examples{

#  this is a DAG where Z is a confounder of the X -> Y relationship
true.model <- 'X ~ .5*Z
               Y ~ .5*X + .5*Z'

# misspecified model omitting Z
fit.model <- 'Y ~ X'

result <- regsim(reps=1000, n=100, true.model, fit.model, targetparm="X",
                 targetval=.5)
result

# visualize the path diagram implied by true.model
plot(result)


}
